Refactor names and folderlocations, update boarditem, add boards page
