+++ src/app/(protected)/_components/Header.tsx
@@ -1,7 +1,7 @@
 import CreateBoardDialog from "@/app/(protected)/_components/CreateBoardDialog";
 import MainNavigation from "@/app/(protected)/_components/MainNavigation";
 import LogoutButton from "@/app/(public)/_components/LogoutButton";
-import Link from "@/components/ui/typography/link";
+import Link from "@/components/typography/link";
 import { HStack } from "@chakra-ui/react";
 
 const Header = () => {
@@ -11,7 +11,7 @@ const Header = () => {
         <Link href="/">Hindsight</Link>
         <MainNavigation />
       </HStack>
-      <HStack gap="1rem">
+      <HStack>
         <CreateBoardDialog />
         <LogoutButton />
       </HStack>
